+++ lua/obsidian/commands/open.lua
@@ -19,8 +19,8 @@ return function(client, data)
       return
     end
   else
-    local cursor_link, _, ref_type = util.cursor_link()
-    if cursor_link ~= nil and ref_type ~= RefTypes.NakedUrl then
+    local cursor_link, _, ref_type = util.parse_cursor_link()
+    if cursor_link ~= nil and ref_type ~= RefTypes.NakedUrl and ref_type ~= RefTypes.FileUrl then
       local note = client:resolve_note(cursor_link)
       if note ~= nil then
         path = assert(client:vault_relative_path(note.path))