+++ lua/obsidian/pickers/_fzf.lua
@@ -105,18 +105,34 @@ FzfPicker.grep = function(self, opts)
   end
 end
 
----@param values string[]
----@param opts { prompt_title: string|?, callback: fun(value: string)|? }|?
+---@param values string[]|obsidian.PickerEntry[]
+---@param opts { prompt_title: string|?, callback: fun(value: any)|? }|?
 ---@diagnostic disable-next-line: unused-local
 FzfPicker.pick = function(self, values, opts)
   opts = opts and opts or {}
 
-  fzf.fzf_exec(values, {
+  ---@type table<string, any>
+  local display_to_value_map = {}
+
+  ---@type string[]
+  local entries = {}
+  for _, value in ipairs(values) do
+    if type(value) == "string" then
+      display_to_value_map[value] = value
+      entries[#entries + 1] = value
+    elseif value.valid ~= false then
+      local display = self:_make_display(value)
+      display_to_value_map[display] = value.value
+      entries[#entries + 1] = display
+    end
+  end
+
+  fzf.fzf_exec(entries, {
     prompt = get_prompt(opts.prompt_title),
     actions = {
       default = function(selected)
         if opts.callback then
-          opts.callback(selected[1])
+          opts.callback(display_to_value_map[selected[1]])
         end
       end,
     },