+++ lua/obsidian/templates.lua
@@ -85,7 +85,7 @@ end
 
 ---Insert a template at the given location.
 ---
----@param name string name of a template in the configured templates folder
+---@param name string name or path of a template in the configured templates folder
 ---@param client obsidian.Client
 ---@param location table a tuple with {bufnr, winnr, row, col}
 M.insert_template = function(name, client, location)
@@ -94,9 +94,29 @@ M.insert_template = function(name, client, location)
     log.err "Templates folder is not defined or does not exist"
     return
   end
-  local buf, win, row, col = unpack(location)
-  local template_path = templates_dir / name
+  local buf, win, row, _ = unpack(location)
   local title = require("obsidian.note").from_buffer(buf, client.dir):display_name()
+
+  ---@type Path
+  local template_path
+  local paths_to_check = { templates_dir / name, Path:new(name) }
+  for _, path in ipairs(paths_to_check) do
+    if path:is_file() then
+      template_path = path
+      break
+    elseif not vim.endswith(tostring(path), ".md") then
+      local path_with_suffix = Path:new(tostring(path) .. ".md")
+      if path_with_suffix:is_file() then
+        template_path = path_with_suffix
+        break
+      end
+    end
+  end
+
+  if template_path == nil then
+    log.err("Template '%s' not found", name)
+    return
+  end
 
   local insert_lines = {}
   local template_file = io.open(tostring(template_path), "r")
@@ -120,13 +140,12 @@ M.insert_template = function(name, client, location)
       end
     end
     template_file:close()
-    table.insert(insert_lines, "")
   else
     log.err("Template file '%s' not found", template_path)
     return
   end
 
-  vim.api.nvim_buf_set_text(buf, row - 1, col, row - 1, col, insert_lines)
+  vim.api.nvim_buf_set_lines(buf, row, row, false, insert_lines)
   local new_cursor_row, _ = unpack(vim.api.nvim_win_get_cursor(win))
   vim.api.nvim_win_set_cursor(0, { new_cursor_row, 0 })