+++ test/obsidian/workspace_spec.lua
@@ -24,7 +24,7 @@ describe("Workspace", function()
   it("should be able to initialize a workspace", function()
     local ws = workspace.new("/tmp/obsidian_test_workspace", { name = "test_workspace" })
     assert.equals("test_workspace", ws.name)
-    assert.equals("/tmp/obsidian_test_workspace", ws.path)
+    assert.equals(vim.fn.resolve "/tmp/obsidian_test_workspace", ws.path)
   end)
 
   it("should be able to initialize from cwd", function()